Team generation moved into scheduler
